refactor: extract ZedImportCard (#3501 Phase 1p)
